+++ component/test/uk/gov/hmrc/apidefinition/ApiDefinitionSpec.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2021 HM Revenue & Customs
+ * Copyright 2024 HM Revenue & Customs
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package component
+package uk.gov.hmrc.definition
 
 import java.util.UUID
 
@@ -24,16 +24,19 @@ import com.github.tomakehurst.wiremock.client.WireMock._
 import play.api.http.HeaderNames.CONTENT_TYPE
 import play.api.http.Status.{NO_CONTENT, OK}
 
-class ApiDefinitionSpec extends ComponentSpec{
+class ApiDefinitionSpec extends ComponentSpec {
 
-  val validDefinition1AsString = """{"serviceName":"calendar","serviceBaseUrl":"http://calendar","name":"Calendar API","description":"My Calendar API","context":"individuals/calendar","versions":[{"version":"1.0","status":"STABLE","access":{"type":"PUBLIC"},"endpoints":[{"uriPattern":"/today","endpointName":"Get Today's Date","method":"GET","authType":"NONE","throttlingTier":"UNLIMITED","queryParameters":[]}],"endpointsEnabled":true,"versionSource":"UNKNOWN"}],"requiresTrust":true,"isTestSupport":false,"categories":["OTHER"]}"""
-  val validDefinition2AsString = """{"serviceName":"calendar","serviceBaseUrl":"http://calendar","name":"Calendar API","description":"My Calendar API","context":"individuals/calendar","versions":[{"version":"1.0","status":"PUBLISHED","endpoints":[{"uriPattern":"/today","endpointName":"Get Today's Date","method":"GET","authType":"NONE","throttlingTier":"UNLIMITED"}]}],"isTestSupport":false,"categories":["OTHER"]}"""
-  
+  val validDefinition1AsString =
+    """{"serviceName":"calendar","serviceBaseUrl":"http://calendar","name":"Calendar API","description":"My Calendar API","context":"individuals/calendar","versions":[{"version":"1.0","status":"STABLE","access":{"type":"PUBLIC"},"endpoints":[{"uriPattern":"/today","endpointName":"Get Today's Date","method":"GET","authType":"NONE","throttlingTier":"UNLIMITED","queryParameters":[]}],"endpointsEnabled":true,"versionSource":"UNKNOWN"}],"requiresTrust":true,"isTestSupport":false,"categories":["OTHER"]}"""
+
+  val validDefinition2AsString =
+    """{"serviceName":"calendar","serviceBaseUrl":"http://calendar","name":"Calendar API","description":"My Calendar API","context":"individuals/calendar","versions":[{"version":"1.0","status":"PUBLISHED","endpoints":[{"uriPattern":"/today","endpointName":"Get Today's Date","method":"GET","authType":"NONE","throttlingTier":"UNLIMITED"}]}],"isTestSupport":false,"categories":["OTHER"]}"""
+
   val validHeaders = List(CONTENT_TYPE -> "application/json")
 
   def stubAWSGateway(): Unit = {
     val expectedRequestId = UUID.randomUUID.toString
-    
+
     stubFor(
       WireMock.put(urlPathEqualTo("/v1/api/individuals--calendar--1.0"))
         .willReturn(
@@ -55,15 +58,15 @@ class ApiDefinitionSpec extends ComponentSpec{
   "POST /api-definition" should {
     "return No Content for valid Json payload" in {
       stubAWSGateway()
-     val response =  post("/api-definition", validDefinition1AsString, validHeaders)
-     response.status shouldBe NO_CONTENT
+      val response = post("/api-definition", validDefinition1AsString, validHeaders)
+      response.status shouldBe NO_CONTENT
     }
 
     "return No Content and be tolerant for json missing values" in {
       stubAWSGateway()
-     val response =  post("/api-definition", validDefinition2AsString, validHeaders)
+      val response = post("/api-definition", validDefinition2AsString, validHeaders)
 
-     response.status shouldBe NO_CONTENT
+      response.status shouldBe NO_CONTENT
     }
   }
 }
